attributes on chat completion
