+++ sentry_sdk/integrations/openai.py
@@ -11,6 +11,7 @@ from sentry_sdk.scope import should_send_default_pii
 from sentry_sdk.utils import (
     capture_internal_exceptions,
     event_from_exception,
+    safe_serialize,
 )
 
 from typing import TYPE_CHECKING
@@ -183,24 +184,50 @@ def _new_chat_completion_common(f, *args, **kwargs):
     )
     span.__enter__()
 
+    # Common attributes
+    set_data_normalized(span, SPANDATA.GEN_AI_SYSTEM, "openai")
+    set_data_normalized(span, SPANDATA.GEN_AI_REQUEST_MODEL, model)
+    set_data_normalized(span, SPANDATA.GEN_AI_OPERATION_NAME, "chat")
+    set_data_normalized(span, SPANDATA.AI_STREAMING, streaming)
+
+    # Optional attributes
+    max_tokens = kwargs.get("max_tokens")
+    if max_tokens is not None:
+        set_data_normalized(span, SPANDATA.GEN_AI_REQUEST_MAX_TOKENS, max_tokens)
+
+    presence_penalty = kwargs.get("presence_penalty")
+    if presence_penalty is not None:
+        set_data_normalized(
+            span, SPANDATA.GEN_AI_REQUEST_PRESENCE_PENALTY, presence_penalty
+        )
+
+    temperature = kwargs.get("temperature")
+    if temperature is not None:
+        set_data_normalized(span, SPANDATA.GEN_AI_REQUEST_TEMPERATURE, temperature)
+
+    top_p = kwargs.get("top_p")
+    if top_p is not None:
+        set_data_normalized(span, SPANDATA.GEN_AI_REQUEST_TOP_P, top_p)
+
     res = yield f, args, kwargs
 
     with capture_internal_exceptions():
         if should_send_default_pii() and integration.include_prompts:
             set_data_normalized(span, SPANDATA.GEN_AI_REQUEST_MESSAGES, messages)
 
-        set_data_normalized(span, SPANDATA.GEN_AI_SYSTEM, "openai")
-        set_data_normalized(span, SPANDATA.GEN_AI_REQUEST_MODEL, model)
-        set_data_normalized(span, SPANDATA.GEN_AI_OPERATION_NAME, "chat")
-        set_data_normalized(span, SPANDATA.AI_STREAMING, streaming)
+        if hasattr(res, "model"):
+            set_data_normalized(span, SPANDATA.GEN_AI_RESPONSE_MODEL, res.model)
 
         if hasattr(res, "choices"):
             if should_send_default_pii() and integration.include_prompts:
-                set_data_normalized(
-                    span,
-                    SPANDATA.GEN_AI_RESPONSE_TEXT,
-                    list(map(lambda x: x.message, res.choices)),
-                )
+                response_text = [choice.message.dict() for choice in res.choices]
+                if len(response_text) > 0:
+                    set_data_normalized(
+                        span,
+                        SPANDATA.GEN_AI_RESPONSE_TEXT,
+                        safe_serialize(response_text),
+                    )
+
             _calculate_token_usage(messages, res, span, None, integration.count_tokens)
             span.__exit__(None, None, None)
         elif hasattr(res, "_iterator"):
